server_code/coap_server/requirements.txt
aiocoap
mysql-connector
